Migrate bot from Bale to Matrix/Element
Replace the Telegram-Bot-API-based transport (pointed at Bale) with a
raw Matrix Client-Server API client and a reaction-based interaction
model, since Matrix has no inline-keyboard/callback-query equivalent.

- New matrix/ package: MatrixClient (OkHttp, access-token auth,
  sync/send/edit/react/join) and MatrixLongPollingBot (sync loop +
  dispatch), replacing TelegramLongPollingBot.
- Menus, roll-call rosters, and confirm/cancel flows now use emoji
  reactions instead of inline keyboards (numbered keycaps for
  per-member rows).
- chatId/userId switched from Long to String across entities,
  services, and models to match Matrix's string-based room/user ids.
- Config now takes matrix.homeserver.url/matrix.access.token instead
  of the Bale bot token/base URL.
- Stop tracking application.properties (it held a live Bale token in
  plaintext) and add application.properties.example as a template.

+++ src/main/kotlin/ir.mhdolatabadi/services/ChatSettingService.kt
@@ -7,7 +7,7 @@ class SettingService(private val emf: EntityManagerFactory) {
     fun isDailyQuestionEnabled(chatId: String): Boolean {
         val entityManager = emf.createEntityManager()
         entityManager.use { em ->
-            val setting = em.find(ChatSetting::class.java, chatId.toLong())
+            val setting = em.find(ChatSetting::class.java, chatId)
             return setting?.dailyQuestionEnabled ?: true
         }
     }
@@ -16,9 +16,9 @@ class SettingService(private val emf: EntityManagerFactory) {
         val em = emf.createEntityManager()
         try {
             em.transaction.begin()
-            val setting = em.find(ChatSetting::class.java, chatId.toLong())
+            val setting = em.find(ChatSetting::class.java, chatId)
             if (setting == null) {
-                val newSetting = ChatSetting(chatId = chatId.toLong(), dailyQuestionEnabled = enabled)
+                val newSetting = ChatSetting(chatId = chatId, dailyQuestionEnabled = enabled)
                 em.persist(newSetting)
             } else {
                 setting.dailyQuestionEnabled = enabled
